Add 'stockfish' commandline switch; move helper functions to utils package.
